feat: refactor autowire
