+++ 2-advanced/dubbo-samples-autowire/dubbo-samples-autowire-consumer/src/main/resources/log4j.properties
