+++ 2-advanced/dubbo-samples-autowire/dubbo-samples-autowire-consumer/src/test/java/org/apache/dubbo/samples/autowire/consumer/AutowireServiceIT.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.dubbo.samples.autowire.consumer;
+
+import org.apache.dubbo.config.annotation.DubboReference;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@SpringBootTest
+@RunWith(SpringRunner.class)
+public class AutowireServiceIT {
+
+    @DubboReference
+    private HelloServiceConsumer helloServiceConsumer;
+
+    @Test
+    public void testSayHello() throws Exception {
+        Assert.assertEquals("HelloServiceConsumer say hello dubbo", helloServiceConsumer.helloServiceConsumerSayHello("dubbo"));
+    }
+
+}
